ShaderLib: Unified glsl.js files.
