--- src/renderers/shaders/ShaderLib/background_frag.glsl.js
@@ -1,16 +1,0 @@
-export default /* glsl */`
-uniform sampler2D t2D;
-
-varying vec2 vUv;
-
-void main() {
-
-	vec4 texColor = texture2D( t2D, vUv );
-
-	gl_FragColor = mapTexelToLinear( texColor );
-
-	#include <tonemapping_fragment>
-	#include <encodings_fragment>
-
-}
-`;
