--- src/renderers/shaders/ShaderLib/background_vert.glsl.js
@@ -1,12 +1,0 @@
-export default /* glsl */`
-varying vec2 vUv;
-uniform mat3 uvTransform;
-
-void main() {
-
-	vUv = ( uvTransform * vec3( uv, 1 ) ).xy;
-
-	gl_Position = vec4( position.xy, 1.0, 1.0 );
-
-}
-`;
