+++ src/renderers/shaders/ShaderLib/cube.glsl.js
@@ -1,4 +1,21 @@
-export default /* glsl */`
+export const vertex = /* glsl */`
+varying vec3 vWorldDirection;
+
+#include <common>
+
+void main() {
+
+	vWorldDirection = transformDirection( position, modelMatrix );
+
+	#include <begin_vertex>
+	#include <project_vertex>
+
+	gl_Position.z = gl_Position.w; // set z to camera.far
+
+}
+`;
+
+export const fragment = /* glsl */`
 #include <envmap_common_pars_fragment>
 uniform float opacity;
