--- src/renderers/shaders/ShaderLib/cube_vert.glsl.js
@@ -1,16 +1,0 @@
-export default /* glsl */`
-varying vec3 vWorldDirection;
-
-#include <common>
-
-void main() {
-
-	vWorldDirection = transformDirection( position, modelMatrix );
-
-	#include <begin_vertex>
-	#include <project_vertex>
-
-	gl_Position.z = gl_Position.w; // set z to camera.far
-
-}
-`;
