--- src/renderers/shaders/ShaderLib/depth_vert.glsl.js
@@ -1,40 +1,0 @@
-export default /* glsl */`
-#include <common>
-#include <uv_pars_vertex>
-#include <displacementmap_pars_vertex>
-#include <morphtarget_pars_vertex>
-#include <skinning_pars_vertex>
-#include <logdepthbuf_pars_vertex>
-#include <clipping_planes_pars_vertex>
-
-// This is used for computing an equivalent of gl_FragCoord.z that is as high precision as possible.
-// Some platforms compute gl_FragCoord at a lower precision which makes the manually computed value better for
-// depth-based postprocessing effects. Reproduced on iPad with A10 processor / iPadOS 13.3.1.
-varying vec2 vHighPrecisionZW;
-
-void main() {
-
-	#include <uv_vertex>
-
-	#include <skinbase_vertex>
-
-	#ifdef USE_DISPLACEMENTMAP
-
-		#include <beginnormal_vertex>
-		#include <morphnormal_vertex>
-		#include <skinnormal_vertex>
-
-	#endif
-
-	#include <begin_vertex>
-	#include <morphtarget_vertex>
-	#include <skinning_vertex>
-	#include <displacementmap_vertex>
-	#include <project_vertex>
-	#include <logdepthbuf_vertex>
-	#include <clipping_planes_vertex>
-
-	vHighPrecisionZW = gl_Position.zw;
-
-}
-`;
