--- src/renderers/shaders/ShaderLib/distanceRGBA_vert.glsl.js
@@ -1,38 +1,0 @@
-export default /* glsl */`
-#define DISTANCE
-
-varying vec3 vWorldPosition;
-
-#include <common>
-#include <uv_pars_vertex>
-#include <displacementmap_pars_vertex>
-#include <morphtarget_pars_vertex>
-#include <skinning_pars_vertex>
-#include <clipping_planes_pars_vertex>
-
-void main() {
-
-	#include <uv_vertex>
-
-	#include <skinbase_vertex>
-
-	#ifdef USE_DISPLACEMENTMAP
-
-		#include <beginnormal_vertex>
-		#include <morphnormal_vertex>
-		#include <skinnormal_vertex>
-
-	#endif
-
-	#include <begin_vertex>
-	#include <morphtarget_vertex>
-	#include <skinning_vertex>
-	#include <displacementmap_vertex>
-	#include <project_vertex>
-	#include <worldpos_vertex>
-	#include <clipping_planes_vertex>
-
-	vWorldPosition = worldPosition.xyz;
-
-}
-`;
