+++ src/renderers/shaders/ShaderLib/equirect.glsl.js
@@ -1,4 +1,19 @@
-export default /* glsl */`
+export const vertex = /* glsl */`
+varying vec3 vWorldDirection;
+
+#include <common>
+
+void main() {
+
+	vWorldDirection = transformDirection( position, modelMatrix );
+
+	#include <begin_vertex>
+	#include <project_vertex>
+
+}
+`;
+
+export const fragment = /* glsl */`
 uniform sampler2D tEquirect;
 
 varying vec3 vWorldDirection;
