--- src/renderers/shaders/ShaderLib/equirect_vert.glsl.js
@@ -1,14 +1,0 @@
-export default /* glsl */`
-varying vec3 vWorldDirection;
-
-#include <common>
-
-void main() {
-
-	vWorldDirection = transformDirection( position, modelMatrix );
-
-	#include <begin_vertex>
-	#include <project_vertex>
-
-}
-`;
