--- src/renderers/shaders/ShaderLib/linedashed_vert.glsl.js
@@ -1,27 +1,0 @@
-export default /* glsl */`
-uniform float scale;
-attribute float lineDistance;
-
-varying float vLineDistance;
-
-#include <common>
-#include <color_pars_vertex>
-#include <fog_pars_vertex>
-#include <morphtarget_pars_vertex>
-#include <logdepthbuf_pars_vertex>
-#include <clipping_planes_pars_vertex>
-
-void main() {
-
-	vLineDistance = scale * lineDistance;
-
-	#include <color_vertex>
-	#include <begin_vertex>
-	#include <morphtarget_vertex>
-	#include <project_vertex>
-	#include <logdepthbuf_vertex>
-	#include <clipping_planes_vertex>
-	#include <fog_vertex>
-
-}
-`;
