--- src/renderers/shaders/ShaderLib/points_vert.glsl.js
@@ -1,35 +1,0 @@
-export default /* glsl */`
-uniform float size;
-uniform float scale;
-
-#include <common>
-#include <color_pars_vertex>
-#include <fog_pars_vertex>
-#include <morphtarget_pars_vertex>
-#include <logdepthbuf_pars_vertex>
-#include <clipping_planes_pars_vertex>
-
-void main() {
-
-	#include <color_vertex>
-	#include <begin_vertex>
-	#include <morphtarget_vertex>
-	#include <project_vertex>
-
-	gl_PointSize = size;
-
-	#ifdef USE_SIZEATTENUATION
-
-		bool isPerspective = isPerspectiveMatrix( projectionMatrix );
-
-		if ( isPerspective ) gl_PointSize *= ( scale / - mvPosition.z );
-
-	#endif
-
-	#include <logdepthbuf_vertex>
-	#include <clipping_planes_vertex>
-	#include <worldpos_vertex>
-	#include <fog_vertex>
-
-}
-`;
