+++ src/renderers/shaders/ShaderLib/shadow.glsl.js
@@ -1,4 +1,4 @@
-export default /* glsl */`
+export const vertex = /* glsl */`
 #include <common>
 #include <fog_pars_vertex>
 #include <morphtarget_pars_vertex>
@@ -21,6 +21,29 @@ void main() {
 	#include <worldpos_vertex>
 	#include <shadowmap_vertex>
 	#include <fog_vertex>
+
+}
+`;
+
+export const fragment = /* glsl */`
+uniform vec3 color;
+uniform float opacity;
+
+#include <common>
+#include <packing>
+#include <fog_pars_fragment>
+#include <bsdfs>
+#include <lights_pars_begin>
+#include <shadowmap_pars_fragment>
+#include <shadowmask_pars_fragment>
+
+void main() {
+
+	gl_FragColor = vec4( color, opacity * ( 1.0 - getShadowMask() ) );
+
+	#include <tonemapping_fragment>
+	#include <encodings_fragment>
+	#include <fog_fragment>
 
 }
 `;
