--- src/renderers/shaders/ShaderLib/shadow_frag.glsl.js
@@ -1,22 +1,0 @@
-export default /* glsl */`
-uniform vec3 color;
-uniform float opacity;
-
-#include <common>
-#include <packing>
-#include <fog_pars_fragment>
-#include <bsdfs>
-#include <lights_pars_begin>
-#include <shadowmap_pars_fragment>
-#include <shadowmask_pars_fragment>
-
-void main() {
-
-	gl_FragColor = vec4( color, opacity * ( 1.0 - getShadowMask() ) );
-
-	#include <tonemapping_fragment>
-	#include <encodings_fragment>
-	#include <fog_fragment>
-
-}
-`;
