+++ src/renderers/shaders/ShaderLib/vsm.glsl.js
@@ -1,4 +1,12 @@
-export default /* glsl */`
+export const vertex = /* glsl */`
+void main() {
+
+	gl_Position = vec4( position, 1.0 );
+
+}
+`;
+
+export const fragment = /* glsl */`
 uniform sampler2D shadow_pass;
 uniform vec2 resolution;
 uniform float radius;
