--- src/renderers/shaders/ShaderLib/vsm_vert.glsl.js
@@ -1,7 +1,0 @@
-export default /* glsl */`
-void main() {
-
-	gl_Position = vec4( position, 1.0 );
-
-}
-`;
